+++ packages/core/src/enqueue_links/enqueue_links.ts
@@ -1,4 +1,5 @@
-import type { BatchAddRequestsResult, Dictionary } from '@crawlee/types';
+import type { Awaitable, BatchAddRequestsResult, Dictionary } from '@crawlee/types';
+import { type RobotsTxtFile } from '@crawlee/utils';
 import ow from 'ow';
 import { getDomain } from 'tldts';
 import type { SetRequired } from 'type-fest';
@@ -16,6 +17,8 @@ import {
     createRequests,
     filterRequestsByPatterns,
 } from './shared';
+
+export type SkippedRequestCallback = (args: { url: string; reason: 'robotsTxt' }) => Awaitable<void>;
 
 export interface EnqueueLinksOptions extends RequestQueueOperationOptions {
     /** Limit the amount of actually enqueued URLs to this number. Useful for testing across the entire crawling scope. */
@@ -158,6 +161,18 @@ export interface EnqueueLinksOptions extends RequestQueueOperationOptions {
      * You can use this option to wait for adding all of them.
      */
     waitForAllRequestsToBeAdded?: boolean;
+
+    /**
+     * RobotsTxtFile instance for the current request that triggered the `enqueueLinks`.
+     * If provided, disallowed URLs will be ignored.
+     */
+    robotsTxtFile?: RobotsTxtFile;
+
+    /**
+     * When a request is skipped for some reason, you can use this callback to act on it.
+     * This is currently fired only for requests skipped based on robots.txt file.
+     */
+    onSkippedRequest?: SkippedRequestCallback;
 }
 
 /**
@@ -256,6 +271,8 @@ export async function enqueueLinks(
         ow.object.exactShape({
             urls: ow.array.ofType(ow.string),
             requestQueue: ow.object.hasKeys('fetchNextRequest', 'addRequest'),
+            robotsTxtFile: ow.optional.object.hasKeys('isAllowed'),
+            onSkippedRequest: ow.optional.function,
             forefront: ow.optional.boolean,
             skipNavigation: ow.optional.boolean,
             limit: ow.optional.number,
@@ -286,6 +303,8 @@ export async function enqueueLinks(
         transformRequestFunction,
         forefront,
         waitForAllRequestsToBeAdded,
+        robotsTxtFile,
+        onSkippedRequest,
     } = options;
 
     const urlExcludePatternObjects: UrlPatternObject[] = [];
@@ -362,6 +381,27 @@ export async function enqueueLinks(
     }
 
     let requestOptions = createRequestOptions(urls, options);
+
+    if (robotsTxtFile) {
+        const skippedRequests: RequestOptions[] = [];
+
+        requestOptions = requestOptions.filter((request) => {
+            if (robotsTxtFile.isAllowed(request.url)) {
+                return true;
+            }
+
+            skippedRequests.push(request);
+            return false;
+        });
+
+        if (onSkippedRequest && skippedRequests.length > 0) {
+            await Promise.all(
+                skippedRequests.map((request) => {
+                    return onSkippedRequest({ url: request.url, reason: 'robotsTxt' });
+                }),
+            );
+        }
+    }
 
     if (transformRequestFunction) {
         requestOptions = requestOptions